Pin the versions of the dependencies.
What this does is to update our dockerfile to use Jammy, and
then install as many dependencies as possible via the system
packages.  We then take the versions of those dependencies and
pin them in requirements.txt to those versions.  This is all
so that we have a consistent set of dependencies that we know
will work for the lifetime of Ubuntu Jammy, without having
to constantly monitor the upstream projects.  We'll update this
again the next time we switch Ubuntu distributions.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>