Apply a constraints file when installing pip packages within image.
This will ensure that target packages are constrained along with their
dependencies. Since we are constraining packages, the --upgrade argument
is useless as without the addition of --upgrade-strategy eager it will
only try to upgrade the packages specified for install and not their
dependencies (which we're also pinning so that isn't the desired
behavior).